fix/ HANDSHAKE CERTIFICATE_VERIFY_FAILED 에러 해결
get->post로 요청이 바뀌면서 certification에러가 나는 듯 하다. 우선
테스트에서 HttpVerify를 true로 하도록 수정해줬다
