+++ test/api/fetch_weather_test.dart
@@ -12,12 +12,14 @@ import 'package:sweater/module/coordi.dart';
 import 'package:sweater/module/forecast.dart';
 import 'package:sweater/module/gender.dart';
 import 'package:sweater/module/user.dart';
+import 'package:sweater/my_http_overrides.dart';
 import 'package:sweater/providers/weather_provider.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 
 main() async {
   dotenv.testLoad(fileInput: File('test/.env').readAsStringSync());
 
+  HttpOverrides.global = MyHttpOverrides();
   test('fetch_weather 통신 테스트', () async {
     int predictMax = 12;
     int nx = 61, ny = 124;
